+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
             android:name=".screens.verification.VerificationActivity"
             android:exported="false" />
         <activity
-            android:name=".screens.main.ProfileActivity"
+            android:name=".screens.profile.ProfileActivity"
             android:exported="false"
             android:theme="@style/Theme.MusicMap.NoActionBar" />
         <activity
